refactor(delta-patcher, delta-json-path-selector): move files
