+++ packages/core/src/delta-json-path-selector/__tests__/json-path.spec.ts
