+++ packages/core/src/delta-json-path-selector/index.ts
@@ -1,0 +1,2 @@
+export * from './json-path';
+export * from './json-selector-matcher';
