+++ packages/core/src/delta-patcher/index.ts
@@ -1,0 +1,1 @@
+export * from './delta-patcher';
