--- packages/genui-renderer/src/server/index.ts
@@ -1,2 +1,0 @@
-export * from '../../../materials/vue-opentiny-vue';
-export * from '../../../materials/angular-opentiny-ng';
